+++ core/pos_move.cpp
@@ -9,10 +9,10 @@ pos_move::pos_move(const position &f, const position &t) :
 
 pos_move::pos_move(const string &iccs)
 {
-    from.file = static_cast<int>(iccs[0] - 'a');
-    from.rank = static_cast<int>(iccs[1] - '0');
-    to.file = static_cast<int>(iccs[2] - 'a');
-    to.rank = static_cast<int>(iccs[3] - '0');
+    from.file = static_cast<int8_t>(iccs[0] - 'a');
+    from.rank = static_cast<int8_t>(iccs[1] - '0');
+    to.file = static_cast<int8_t>(iccs[2] - 'a');
+    to.rank = static_cast<int8_t>(iccs[3] - '0');
 }
 
 string pos_move::get_iccs() const
